### 2021-04-30 #3229 @whedon
[PRE REVIEW]: HARDy: Handling Arbitrary Recognition of Data in Python
**Submitting author:** @MariaPoliti (<a href="http://orcid.org/0000-0002-5815-3371">Maria Politi</a>)
**Repository:** <a href="https://github.com/EISy-as-Py/hardy" target ="_blank">https://github.com/EISy-as-Py/hardy</a>
**Version:** v1.0
