+++ joss-paper/paper.md
@@ -48,7 +48,7 @@ bibliography: paper.bib
 - _Recognition_: Machine Learning module that applies user defined hyperparameter search space for training and evaluation of model
 - _Data-Reporting_: Imports result of machine learning models and reports it into dataframes and plots
 
-# Background
+# Statement of Need
 
 High Throughput Experimentation (HTE) and High Throughput Testing (HTT) have exponentially increased the volume of experimental data available to scientists. One of the major bottlenecks in their implementation is the data analysis. The need for autonomous binning and classification has seen an increase in the employment of machine learning approaches in discovery of catalysts, energy materials and process parameters for design of experiment [@williams2019enabling; @becker2019low]. However, these solutions rely on specific sets of hyperparameters for their machine learning models to achieve the desired purpose. Furthermore, numerical data from experimental characterization of materials carries diversity in both features and magnitude. These features are traditionally extracted using deterministic models based on empirical relationships between variables of the process under investigation. As an example, X-ray diffraction (XRD) data is easier to characterize in linear form as compared to small angle X-ray scattering data, which requires transformation of axis to log-log scale.
 
