### 2021-04-30 3b9c5bc
Added doi information in bibliography
