+++ joss-paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={157--165},
   year={2019},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.chemmater.9b03043.s001}
 }
 
 @article{becker2019low,
@@ -17,7 +18,8 @@
   number={22},
   pages={1900555},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/aenm.201900555}
 }
 
 @book{hutter2019automated,
@@ -47,7 +49,8 @@
   booktitle={2015 IEEE international conference on data science and advanced analytics (DSAA)},
   pages={1--10},
   year={2015},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/dsaa.2015.7344858}
 }
 
 @inproceedings{nargesian2017learning,
@@ -55,7 +58,8 @@
   author={Nargesian, Fatemeh and Samulowitz, Horst and Khurana, Udayan and Khalil, Elias B and Turaga, Deepak S},
   booktitle={IJCAI},
   pages={2529--2535},
-  year={2017}
+  year={2017},
+  doi={10.24963/ijcai.2017/352}
 }
 
 @article{ArchibaldRichardK2020Caas,
@@ -72,4 +76,5 @@ language = {eng},
 address = {5 Abbey Square, Chester, Cheshire CH1 2HU, England},
 author = {Archibald, Richard K and Doucet, Mathieu and Johnston, Travis and Young, Steven R and Yang, Erika and Heller, William T},
 organization = {Oak Ridge National Lab. (ORNL), Oak Ridge, TN (United States)},
+doi={10.1107/s1600576720000552}
 }
