### 2021-05-01 #3229 @MariaPoliti
@gkthiruvathukal <https://github.com/gkthiruvathukal>,

Thank you so much for volunteering to be the editor for our submission.
Here is a list of reviewer who we think could be a good fit:

- yxoos,
- rmsare
- tacaswell

I am happy to provide other names if needed.
Best,
Maria & Moeez

On Fri, Apr 30, 2021 at 9:02 PM George K. Thiruvathukal <
***@***.***> wrote:

> @MariaPoliti <https://github.com/MariaPoliti> Can you please take a look
> at the list of people who can review for JOSS and suggest 2-4 of them? This
> will greatly help me to get the review process started.
>
