### 2021-06-28 #3229 @yxoos
Sorry, I am so busy with work, cannot do it.
______________________
Amir Hassanzadeh

Ph.D. Candidate
Chester F. Carlson Center for Imaging Science
Rochester Institute of Technology, NY, 14623, USA
Email:     ***@***.***
Tel:         585-771-0680

On Mon, Jun 28, 2021 at 8:48 AM George K. Thiruvathukal <
***@***.***> wrote:

> @yxoos <https://github.com/yxoos> Would you be willing to contribute a
> review for this JOSS submission? I need two reviewers, so I am hoping you
> and @tacaswell <https://github.com/tacaswell> can help me out!
>
