### 2021-08-24 EISy-as-Py/hardy#6 @andrewtarzia
[BUG] : Example notebook link in README does not work.
**Describe the bug**
The link the in the readme to an example notebook does not work.

 - Browser: chome
