### 2021-08-24 EISy-as-Py/hardy#7 @andrewtarzia
[Discussion] : Timings for trainings in the JOSS paper
** Your Use-Case **
Hello, I am reviewing this package for JOSS and also imagine a use-case in our lab (at Imperial College) on classifying different spectra in high-throughput automated synthesis.

This discussion, or issue, is just to get a guide-line for the compute time and requirements to an experimental lab. I.e. how long did the training take for the set in the JOSS paper?
