### 2021-08-24 #3229 @andrewtarzia
Started to review this. Lacks usage introduction. If there is a compiled documentation, it is not accessible.

Added an issue for the example notebook not working: https://github.com/EISy-as-Py/hardy/issues/6

Interested in training timing: https://github.com/EISy-as-Py/hardy/issues/7
