### 2021-08-24 ec7308e
Added the ReadTheDocs link in the Readme.md in response to the Issue # 6
