### 2021-08-24 EISy-as-Py/hardy#6 @MariaPoliti
@andrewtarzia The link for example notebook has been fixed. The link for compiled documentation has also been added in Readme.md. I hope this helps.
