### 2021-08-30 686be6e
Added the dependencies for Hardy in setup.py to enable pip install
