### 2021-08-30 5559db0
bumped the PyYAML to version 5.4.0 to overcome the security log issued by dependabot on github. The change has been tested by python3.7 -m unittest discover on Mac
