+++ setup.py
@@ -48,7 +48,7 @@ setup(name='hardy', version=1.0,
           "scikit-image==0.17.2",
           "opencv-python",
           "grpcio==1.16.1",
-          "pyyaml==5.3.1",
+          "pyyaml==5.4",
           "scipy==1.4.1",
           "python-dateutil==2.8.1",
           "pytz==2020.1",
