### 2021-08-31 ab6d8c5
Increasing the travis timeout wait since the default time_out is 10 minutes which is not enough for resolution of dependencies
