+++ .travis.yml
@@ -25,6 +25,7 @@ before_install:
 
 # list of commands to run to setup the environment
 install:
+    - travis_wait 30 mvn install
     - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION --file environment.yml
     - conda activate test-environment
 #    - pip install --no-deps keras-tuner
