### 2021-08-31 d970b6a
Adding the travis wait for 30 minutes to the conda environment creation to avoid timeout
