+++ .travis.yml
@@ -25,8 +25,7 @@ before_install:
 
 # list of commands to run to setup the environment
 install:
-    - travis_wait 30 mvn install
-    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION --file environment.yml
+    - travis wait 30 conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION --file environment.yml
     - conda activate test-environment
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
