### 2021-08-31 0b56a75
fixing the typographical error on the travis_wait
