### 2021-08-31 31788a0
removing the travis_wait since the script executes but still doesn't resolve the environment; would be checking travis bandwidth
