### 2021-08-31 b31a826
Using pip to install hardy to avoid the conda dependencies error
