+++ .travis.yml
@@ -25,10 +25,11 @@ before_install:
 
 # list of commands to run to setup the environment
 install:
-    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION --file environment.yml
+    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
     - conda activate test-environment
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
+    - python setup.py install
 
 # a list of commands to run before the main script
 before_script:
