### 2021-08-31 b96a169
removing the conda environment create to test the system on pip
