+++ .travis.yml
@@ -25,8 +25,8 @@ before_install:
 
 # list of commands to run to setup the environment
 install:
-    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
-    - conda activate test-environment
+#    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
+#    - conda activate test-environment
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
     - python setup.py install
