### 2021-08-31 fe412aa
adding the miniconda version for python3.7
