### 2021-08-31 9ecb7c1
added the -q tag back to the conda create
