### 2021-08-31 bc91b68
Adding the pip install again. Also removed the environment.yml
