+++ setup.py
@@ -47,7 +47,7 @@ setup(name='hardy', version=1.0,
           "plotly==4.8.1",
           "scikit-image==0.17.2",
           "opencv-python",
-          "grpcio==1.16.1",
+          "grpcio==1.24.3",
           "pyyaml==5.4",
           "scipy==1.4.1",
           "python-dateutil==2.8.1",
