### 2021-10-16 #3829 @whedon
[REVIEW]: HARDy: Handling Arbitrary Recognition of Data in Python
**Submitting author:** @MariaPoliti (<a href="http://orcid.org/0000-0002-5815-3371">Maria Politi</a>)
**Repository:** https://github.com/EISy-as-Py/hardy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
