### 2021-10-25 EISy-as-Py/hardy#10 @andrewtarzia
[Discussion] : Approach to installation.
Further to #9, I found that the installation instructions (README, and the two pages at https://hardy.readthedocs.io/en/latest/index.html) were not entirely consistent.
