### 2021-10-25 EISy-as-Py/hardy#11 @andrewtarzia
[Discussion] : Provide example trained models.
Hello!

** Your Use-Case **
Reviewing this software is difficult for me because I do not currently have the infrastructure to complete the model training and I am struggling to get started with the software based on the available examples. #3 suggests there are some more tutorials/examples coming - will these include trained models to perform the data reporting?
