### 2021-10-28 d322986
Bumping the requirements of python to 3.7 from 3.6 since conda installation fails due to dependency error
