### 2021-10-28 ce1a00d
Adding PyYAML to the requirements.txt to resolve the issue with the sphinx readthedocs
