### 2021-10-28 1468b3a
Adding sklearn to requirement.txt to resolve the failing docs, PyYAML issue resolved
