+++ doc/requirements.txt
@@ -10,3 +10,4 @@ plotly
 nbsphinx
 numpydoc
 PyYAML
+scikit-learn
