### 2021-10-28 fca803a
resolving issues with readthedocs by removing warnings
