### 2021-10-28 a31868d
removing the typographical errors to resolve issues with readthedocs
