### 2021-10-28 b56605e
resolving flake8 issue in the arbitrage.py
