+++ doc/requirements.txt
@@ -11,3 +11,4 @@ nbsphinx
 numpydoc
 PyYAML
 scikit-learn
+docutils==0.17.1
