### 2021-10-30 EISy-as-Py/hardy#9 @amoeez
@andrewtarzia We started having troubles with Conda install on python 3.6 couple of months back (appeared in Travis CI). We traced changes in packages version upgrades, however, it didn't produce any meaningful solution. The only option we currently have is to the keep the python version restricted to 3.7 only.
I hope this will help. The change in dependency requirement has been added to the Readme and the documentations.
