### 2021-10-30 8d59fa4
MChanging the docs to comply the readme.md for installation instructions. Adding llink to installation page on readme
