### 2021-10-30 b00ebd9
cresolving the issue with weblink of installation page in docs and correcting the environment file code view on index.rst
