### 2021-10-31 3262a9b
Making the installation instructions consistent to resolve issue # 10
