+++ doc/installation.rst
@@ -12,13 +12,17 @@ The easiest way to install :code:`HARDy` is using :code:`conda`::
 Installation using Git
 ----------------------
 :code:`HARDy` can also be installed using Git. Currently version 1.0
-is the most stable version. To install version 1.0, run the following
-commands in the terminal::
+is the most stable version. To install version v1.0, follow the following
+steps::
 
+* In your terminal, run::
     git clone https://github.com/EISy-as-Py/hardy.git
-    cd HARDy
+* Change the directory to hardy root directory, by running::
+    cd hardy
+* Run::
     git checkout v1.0
-    python setup.py install
+* Run::
+    python setup.py install</code>
 
 To check installation run the following command in your terminal::
     
