### 2021-10-31 9fb861d
fixing some typographical errors on the index and installation documents page
