### 2021-11-02 EISy-as-Py/hardy#10 @amoeez
@andrewtarzia thank you for pointing that out. I hope the installation instructions appear consistent now.
