### 2021-11-07 EISy-as-Py/hardy#9 @andrewtarzia
Ok, excellent. I am using Python 3.7 now through Conda.
