### 2021-11-08 #3829 @andrewtarzia
> I have read the manuscript and very excited to review and test out this software!
>
> Created some issues about getting started and installation, which has not been smooth for me - although I acknowledge that it could be because I have not used TensorFlow before and I am not used to its installation: [EISy-as-Py/hardy#9](https://github.com/EISy-as-Py/hardy/issues/9) [EISy-as-Py/hardy#10](https://github.com/EISy-as-Py/hardy/issues/10) [EISy-as-Py/hardy#11](https://github.com/EISy-as-Py/hardy/issues/11)
>
> Comments on the manuscript:
>
> * The figures are a little unclear. Some explanation of Fig 1 would be great.
> * Explanation of the left-most column in Fig 2 and the distinction between symbols in Figure 3 are necessary.

The above has mostly been answered. Still awaiting a way to test parts of Hardy that do not require the costly training that I cannot perform (issue 11). At this point, it is still unclear how to use Hardy after training a model that can make the classifications.

The checklist above highlights what is missing.
