### 2021-11-09 EISy-as-Py/hardy#11 @amoeez
@andrewtarzia Yes, we plan to include trained models. For now, we've created a temporary Google Colab notebook (https://colab.research.google.com/drive/1XzlhTitEmeSmhieP1njmsHPk24weFGiL?usp=sharing) until we find an appropriate location to host this data. The notebook includes the performance comparison of trained models on both RGB and Cartesian Coordinate representation for Small Angle X-Ray Scattering (SAXS) simulated data. Please note that you'll be downloading these trained models into your own drive, so these should be accessible to you. I hope this helps. Please reach out if there are any other questions.
