### 2021-11-10 2f02fda
Bump tensorflow from 2.2.0 to 2.5.2
Bumps [tensorflow](https://github.com/tensorflow/tensorflow) from 2.2.0 to 2.5.2.
- [Release notes](https://github.com/tensorflow/tensorflow/releases)
- [Changelog](https://github.com/tensorflow/tensorflow/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorflow/compare/v2.2.0...v2.5.2)

---
updated-dependencies:
- dependency-name: tensorflow
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
