### 2021-11-11 EISy-as-Py/hardy#11 @andrewtarzia
@amoeezuw are there examples for making predictions on a new piece of data?
