### 2021-11-13 1830ad1
Adding the feature to remove dependency of test_data on the labels generated
