### 2021-11-13 d16b1f8
Added the extension of h5 to the model save in tuner since the pb extension is generating incorrect outputs
