### 2021-11-14 9844e06
added the module for making predictions on new data set in reporting
