### 2021-11-17 d46f0ac
Updated images and added more descriptive captions
