+++ joss-paper/paper.md
@@ -62,19 +62,19 @@ One of the most widely applied strategy to enhance the performance of machine le
 
 The python-based package `HARDy` is a modularly structured package which classifies data using 2D convolutional neural networks. A schematic for the package can be found in figure 1.
 
-![Information flow for HARDy](./images/HARDy_diagram.png)
+![Workflow schematics for HARDy.  After the data files _(.csv)_ are loaded, the transformations indicated in the _configuration file (.yaml)_ will be applied to the indicated variables. The configuration file will also contain information on which axis (x/y for Cartesian Representation or RGB channels for color images) to use for the loaded data. Once the images are generated, they will be used as input for a Convolutional Neural Network (CNN) for the data classification step. Another _configuration file (.yaml)_ is used to indicate the hyperparameters and network structure of the CNN or the hyperparameter space to use for the tuning session. Each combination of transformations will be used to train and test its own CNN/CNN tuning session. Finally, the data reporting will provide a final summary of the best classification accuracy for each transformation tested in the session, as well as the final trained model.](./images/HARDy_diagram.png)
 
 The package was tested on a set of simulated Small Angle Scattering (SAS) data to be classified into four different particle models: spherical, ellipsoidal, cylindrical and core-shell spherical. A total of ten thousand files were generated for each model. The data was generated using \textit{sasmodels}. The geometrical and physical parameters used to obtain each spectrum were taken from a published work discussing a similar classification task [@ArchibaldRichardK2020Caas]. The name of each SAS model was used as label for the data, allowing for further validation of the test set results. These models were selected as they present similar parameters and data features, which at times make it challenging to distinguish between them.
 
-![Summary table of few transformations using cartesian coordinate and RGB representation along with the respective fittings. The final accuracies are also provided. The original data label corresponds to the icon on the left-most graph, whereas the icon under each fit correspond to the model prediction.](./images/transformation_run_example.png)
+![Summary table of few transformations visualized using cartesian coordinate and RGB representation along with the respective fittings. The left-most column shows the transformations applied to the data: the scattering vector _q_ and scattering intensity _I(q)_. The final model accuracies are also provided. The original data label corresponds to the icon on the left-most graph, whereas the icon under each fit correspond to the model prediction.](./images/transformation_run_example.png)
 
 First, the pre-labelled data was loaded. A subset of the files, three thousand files in total, was identified as the testing set. All the ML models initialized in the same code run were validated using the same testing set. A user-provided list of transformations, inputted through a configuration file, was then applied to the data. Different trials can be specified, so that multiple sets of transformations can be investigated. Both Cartesian and RGB plots representations were compared. The latter visualization option was obtained by encoding the data into the pixel values of each channel composing a color image, for a total of six-channels available (i.e., 3 RGB channels in horizontal/vertical orthogonal directions).
 
-The data was then fed into a convolutional neural network, whose hyperparameters and structure were defined using another configuration file. Alternatively, it is also possible to train multiple classifiers for a single transformation trial through the use of a tuner, by instead providing a hyperparameter space and a search method. The classification results, as well as the best performing trained model were saved for each transformation run. The package also allows to visually compare, through parallel coordinates plots (see documentation), the performance of each transformation. Figure 2 shows a summary of few runs comparing the two visualization strategies and their best performing model accuracies. 
[…]
