### 2021-11-18 4894e6e
Added UW Hyak acknowledgement to JOSS paper
