### 2021-11-18 bd746ed
added manual numpy installation to the repository
