+++ .travis.yml
@@ -30,6 +30,7 @@ install:
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
     - pip install .
+    - pip install numpy==1.18.1
 
 # a list of commands to run before the main script
 before_script:
