### 2021-11-18 96b0bc9
checking if travis can discover numpy installed on system, also removed python36 from travis
