### 2021-11-18 cc4452c
added quotes on the python command to check numpy detected by python
