+++ .travis.yml
@@ -29,10 +29,11 @@ install:
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
     - pip install .
+    - python -c "import pandas"
 
 # a list of commands to run before the main script
 before_script:
-    - python -c "import numpy"
+    - python -c "import pandas"
     - flake8 hardy
 
 # the actual commands to run
