### 2021-11-18 1b4730b
using conda environment to install the package rather than pip
