+++ .travis.yml
@@ -24,16 +24,17 @@ before_install:
 
 # list of commands to run to setup the environment
 install:
-#    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
-#   - conda activate test-environment
+    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
+    - conda activate test-environment
+    - conda install install -c pozzorg hardy
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
-    - pip install .
-    - python -c "import pandas"
+#    - pip install .
+#    - python -c "import pandas"
 
 # a list of commands to run before the main script
 before_script:
-    - python -c "import pandas"
+    - conda activate test_environment
     - flake8 hardy
 
 # the actual commands to run
