### 2021-11-18 daac392
reverting back to the pip installation
