+++ .travis.yml
@@ -26,10 +26,10 @@ before_install:
 install:
     - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
     - conda activate test-environment
-    - conda install install -c pozzorg hardy
+#    - conda install install -c pozzorg hardy
 #    - pip install --no-deps keras-tuner
     - conda install --yes coverage coveralls flake8
-#    - pip install .
+    - pip install .
 #    - python -c "import pandas"
 
 # a list of commands to run before the main script
