### 2021-11-18 d7b3f22
removing the python version from conda env create
