+++ .travis.yml
@@ -24,7 +24,8 @@ before_install:
 
 # list of commands to run to setup the environment
 install:
-    - conda env create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
+    - conda env create -q -n test-environment
+#   python=$TRAVIS_PYTHON_VERSION
     - conda activate test-environment
 #    - conda install install -c pozzorg hardy
 #    - pip install --no-deps keras-tuner
