### 2021-11-18 bafb6f7
reverting back all the settings and pip installing everything
