### 2021-11-18 a098d44
solved the travis issue by switching the coverage coveralls and flake8 intallation from before hardy installation to after hardy installation, moreover, removing any conda installed packages
