### 2021-11-18 fbe3791
removed the python 36 version from travis since it causes compatability issues
