### 2021-11-18 EISy-as-Py/hardy#11 @amoeez
@andrewtarzia please find the example for a new data set using trained model on following google colab link (https://colab.research.google.com/drive/1zIIDerpRz28VOp5KQXyS6LkxbJZR5cqB?usp=sharing). The tutorial has also been added in the documentation. We found a bug while reviewing this which didn't allowed us to use the previously trained models. It has now been resolved and you may need to reinstall the HARDy if you intend to use it on the local machine. Please let me know if you've further questions. Thank you
