### 2021-11-22 0b5d51c
Added the community guidelines to the readme.md
