+++ README.md
@@ -94,6 +94,17 @@ Our vision of the final product:
  * __recognition.py__      :  Setup, training and testing of single convolutional neural network (CNN) or hyperparameters optimization for CNNs.
  * __data_reporting.py__   :  Output and reporting of any/all results. Tabular summary of runs, visual performance comparison, as well as parallel coordinate plots and feature maps
 
+ ## Community Guidlines:
+ We welcome contributors to extend the functionalities, generate feature requests and report bugs in HARDy
+ 
+ ### Feature Request:
+ If you would like to suggest a feature or start a discussion on possible extension of HARDy, please <a href="https://github.com/ECSHackWeek/impedance.py/issues/new/choose">raise an issue</a>
+ 
+ ### Bug Report:
+ If you would like to report a bug, the link to raise an issue is <a href="https://github.com/ECSHackWeek/impedance.py/issues/new/choose">Here</a>
+ 
+ ### Contributions:
+ If you would to contribute to HARDy, you can fork the repository, add your contribution and generate a pull request.
 
  ## Acknowledgment
 
