### 2021-11-22 EISy-as-Py/hardy#11 @andrewtarzia
That notebook works! Thank you. I think this tool is very useful but the usability really comes from good examples like this that guide a new user through the process for this test case.
