### 2021-11-22 #3829 @andrewtarzia
I have completed my review and recommend acceptance of this software paper, the work is potentially very useful, although some checklist items above remain unticked for me, which I highlight here:

> Google colab notebooks have been developed throughout this review process to answer issues I have sent that show the analysis and use of models from hardy. These guides should be formalised and made easy to find from the README and docs.

> This goes with the issue above, I think the documentation is scattered in different forms (examples, documentation and READMEs) and is difficult to grasp. Additionally, many functions/methods lack complete documentation (e.g., `data_wrapper` in `run_hardy.py`)

- [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

> The writing is good and the manuscript quality is overall good, just some minor comments I highlighted earlier:
> * The figures are a little unclear. Some explanation of Fig 1 would be great.
> * Explanation of the left most column in Fig 2 and the distinction between symbols in Figure 3 are necessary.
