### 2021-11-22 8e32a3b
fixing a typographical error in the readme.md in the community guidlines section
