+++ CONTRIBUTIONS.md
@@ -1,0 +1,3 @@
+# Welcome to the guide for making contributions to HARDy
+
+Please report all the bugs in HARDy modules and mistakes in documentation by creating an <a href=https://github.com/EISy-as-Py/hardy/issues/new/choose>issue</a>
