+++ README.md
@@ -95,13 +95,13 @@ Our vision of the final product:
  * __data_reporting.py__   :  Output and reporting of any/all results. Tabular summary of runs, visual performance comparison, as well as parallel coordinate plots and feature maps
 
  ## Community Guidlines:
- We welcome contributors to extend the functionalities, generate feature requests and report bugs in HARDy
+ We welcome the members of open-source communtiy to extend the functionalities of HARDy, submit feature requests and report bugs.
  
  ### Feature Request:
- If you would like to suggest a feature or start a discussion on possible extension of HARDy, please <a href="https://github.com/ECSHackWeek/impedance.py/issues/new/choose">raise an issue</a>
+ If you would like to suggest a feature or start a discussion on possible extension of HARDy, please <a href="https://github.com/EISy-as-Py/hardy/issues/new/choose">raise an issue</a>
  
  ### Bug Report:
- If you would like to report a bug, the link to raise an issue is <a href="https://github.com/ECSHackWeek/impedance.py/issues/new/choose">Here</a>
+ If you would like to report a bug, the link to raise an issue is <a href="https://github.com/EISy-as-Py/hardy/issues/new/choose">Here</a>
  
  ### Contributions:
  If you would to contribute to HARDy, you can fork the repository, add your contribution and generate a pull request.
