### 2021-11-22 ad8a26d
Completed the contributions.md, added its link on the community guidelines on readme.md. Fixed some typographical errors on the readme.md
