### 2021-11-27 12595ba
adding math notations for the transformations in the docs
