### 2021-11-27 9263ac4
Adding the mathematics part for transformations in the docstrings
