+++ hardy/arbitrage/transformations.py
@@ -15,7 +15,7 @@ def raw(raw_array):
     -------
     raw_array: numpy.array
                array representing data values
-    
+
     Notes
     -----
     .. math::
@@ -35,12 +35,12 @@ def exp(raw_array):
     ----------
     raw_array: numpy.array
                array representing data values
-    
+
     Returns
     -------
     exp_array: numpy.array
                array representing the exponentials of data values
-    
+
     Notes
     -----
     .. math::
@@ -63,6 +63,11 @@ def nlog(raw_array):
     -------
     log_array: np.ndarray
                natural log values of each element in the input array
+
+    Notes
+    -----
+    .. math::
+        Z = \\ln{Z}
     '''
 
     # NOTE: All Elements in array MUST be Positive!?
@@ -83,6 +88,10 @@ def log10(raw_array):
     -------
     log_array: np.ndarray
                natural log values of each element in the input array
+    Notes
+    -----
+    .. math::
+        Z = \\lg{Z}
     '''
 
     # NOTE: All Elements in array MUST be Positive!?
@@ -103,6 +112,11 @@ def reciprocal(raw_array):
     -------
     reciprocal_array: np.ndarray
                reciprocal values of each element in the input array
+
+    Notes
+    -----
+    .. math::
+        Z = \\frac{1}{Z}
     '''
     reciprocal_array = np.reciprocal(raw_array)
     return reciprocal_array
@@ -119,6 +133,11 @@ def cumsum(raw_array):
     -------
     cumsum _array: np.ndarray
                cumulative sum of values in the input array
+
+    Notes
+    -----
+    .. math::
+        Z = \[ [Z_1, Z_1+Z_2, Z_1+Z_2+Z_3, Z_1+...+Z_n] \]
     '''
     cumsum_array = np.cumsum(raw_array)
     return cumsum_array
@@ -137,6 +156,15 @@ def derivative_1d(raw_array, spacing=0):
     -------
     derivative_array: np.ndarray
                       array representing gradient at each datapoint
+
+    Notes
+    -----
+    .. math::
+        f_{(0)}^{(1)} = \frac{f(x+1)-f(x-1)}{2h}
+
+    where
+    .. math::
+        h = \[ [0, 1, 2, 3, ..., size-1] \]
     '''
 
     if spacing == 0:
