### 2021-11-28 0e74b67
Added all the math notations for the transformation in the transformation.py
