### 2021-11-28 ed58893
Added the getting started notebook in the docs
