### 2021-11-28 c7d13e3
Added the arguments for the hardy_main in the getting started folder
