### 2021-11-29 a99bdad
Fixing the errors in getting started documentations
