### 2021-12-01 9d1a7f0
modified the gstarted.rst in the docs
