### 2021-12-01 c16c9c1
made change to the gstarted.rst
