### 2021-12-01 1dccd12
Aligned the image to the center at gstarted.rst
