### 2021-12-01 d30a3e8
Added the advanced functionalities documentation
