### 2021-12-03 8b1c736
Updated figure 1 caption and removed border from figure 2
