### 2021-12-05 574b15b
Configuring the example of training on new data set
