+++ doc/advanced.rst
@@ -36,7 +36,7 @@ These transformation folders contains model validation results, best tuned model
 for a particular transformation and model hyperparameter details.
 
 Further Contents
-================
+----------------
 
 .. toctree::
     :maxdepth: 1
