### 2021-12-09 41c3724
Added the description of some parameters to be assigned for hardy run
