+++ doc/gstarted.rst
@@ -29,7 +29,8 @@ Data Preparation
 :code:`HARDy` is configured to input :code:`.csv` files only. Before
 starting your :code:`HARDy` run, make sure the data files are only in
 :code:`.csv` format. Moreover, the :code:`.csv` files must have a header
-of same length. 
+of same length. Another parameter to keep in consideration is that all the
+data files must have same number of data rows.
 
 The wrapper function, :code:`run_hardy`, takes care of all the numerical
 and visual transformations along with hyperparameter tuning and CNN runs.
@@ -56,13 +57,52 @@ Defining the path to tuner or CNN configuration::
 
     classifier_config_path = './hardy/recognition/'
 
+Setting up parameters
+---------------------
+The wrapper function of :code:`HARDy (hardy-main)` is capable to intake a large number
+of user defined inputs. However, some important parameters that should be defined before
+the runs are described here:
+
+    * scale: The RGB images generated by :code:`HARDy` are NxN dimensions, where N
+      is the length of data rows in the csv files. It was observed that if the data
+      has 500 rows, the RGB image would have 500x500 dimension. This exponentially
+      increases the demand for RAM. To overcome this issue, :code:`scale` argument
+      has been introduced to scale down the images. It uses :code:`scipy.misc.imresize`
+      to scale down the image. :code:`scale` represents a fractional value for example
+      0.1, 0.2.. 1.0
+    * target_size: target size is the product of :code:`scale` and :code:`N`. It is
+      tuple of shape 1x2. For example, (100, 100)
+    * num_test_files_class: this is an integer value representing the number of files
+      per category/class. This number of test of files will be seperated out from the
+      training and only be utilized during final validation of model. The files are
+      randomly selected. To control which data files are used, main function is capable
+      of taking :code:`seed` as the input.
+    * classes: list of classes/categories in which data is categorized. Remembering the
+      order of classes is important since this order is utilized for classfiying the
+      new data inputs for a trained.
+    * iterator_mode: Since :code:`Keras` is capable of using either numpy array
+      representation of data as well as image data, this string value controls which
+      iterator to be used. If the value is :code:`arrays`, the RGB data representation
+      will be in numpy array format and :code:`Keras` will use numpy iterator. If value
+      other than the :code:`arrays`, PNG images will be generated, saved on disk and will
+      be fed into the :code:`Keras` using a directory iterator.
+    * n_threads: This parameter controls the parallel processing of transformations. Since
+      transformation configuration file may include saveral transformations, this parameter
+      can assign each transformation to each n_thread available. A value of 6 would means
+      each of the 6 threads will be utilized to run 6 transformations in parallel. The
+      :code:`n_thread` parameter doesn't control the threads used for traning and tuning of
+      neural networks. The neural network multiprocessing is either controlled through CUDA
+      or default tensforflow operations.
+
+Other arguments that can be given as input of :code:`hardy_main` function are described below
+
 Executing hardy run
 -------------------
 The following code starts the numerical and visual transformations along with the
 hyperparameter tuning session::
 
     run.hardy_main(raw_data_path, tform_config_path, classifier_config_path, batch_size=64,
-    scale=0.2, num_test_files_class=750, target_size=(500, 500), iterator_mode='arrays',
+    scale=0.2, num_test_files_class=750, target_size=(100, 100), iterator_mode='arrays',
     classifier='tuner', n_threads=1, classes=['class_1', 'class_2', 'class_3'],
     project_name='my_project_name')
 
