### 2021-12-10 7219134
Added the data reporting notebook in the documentations and examples
