### 2021-12-10 27b6cff
Changed the theme to test the plotly images
