### 2021-12-10 35483d4
added cloud-sptheme in the requirements.txt
