+++ doc/requirements.txt
@@ -12,3 +12,4 @@ numpydoc
 PyYAML
 scikit-learn
 docutils==0.17.1
+cloud-sptheme
