### 2021-12-10 2f58356
reverting back to rtd theme to see if plotly works
