+++ doc/conf.py
@@ -92,7 +92,7 @@ todo_include_todos = True
 # a list of builtin themes.
 #
 
-html_theme = 'cloud'
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
