+++ doc/requirements.txt
@@ -12,4 +12,4 @@ numpydoc
 PyYAML
 scikit-learn
 docutils==0.17.1
-cloud-sptheme
+sphinx_rtd_theme==0.5.0rc1
