+++ doc/conf.py
@@ -101,7 +101,7 @@ todo_include_todos = True
 # a list of builtin themes.
 #
 
-html_theme = 'renku'
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
