### 2021-12-14 92da63c
Added the guide for using numpy array iterator vs directory iterator
