### 2021-12-15 8afccac
added the kfold cross validation notebook in the documentations
