### 2021-12-15 50c2938
Added the hyperlink for k-fold cross validation ntoebook in advanced documentation webpage
