### 2021-12-17 496beae
Feature mapping has been modified to build maps for dynamics model layer num, and outputs, the tests have been modified accordingly. moreover, the model build is being imported from the tensorflow.keras instead of keras to resolve the issue the conv2d having no length
