+++ hardy/test/test_cnn.py
@@ -1,6 +1,5 @@
 import keras
 import os
-import pickle
 import shutil
 import unittest
 
@@ -9,6 +8,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from hardy.recognition import cnn
+from hardy.recognition import save_load_model
 from hardy.handling.to_catalogue import learning_set, test_set
 from keras.preprocessing.image import load_img, img_to_array
 
@@ -65,11 +65,11 @@ class TestSimulationTools(unittest.TestCase):
         assert isinstance(report, str), 'the report should be a string'
 
     def test_feature_maps(self):
-        with open('./hardy/test/test_model/test_model.sav', 'rb') as file:
-            model = pickle.load(file)
-        image_name = '200504-0132_sim_one_current_noise_log-freq.png'
-        image_path = './hardy/test/test_image/class_1/'
-        last_layer = cnn.feature_map(image_name, model, 2, 80,
+        model = save_load_model('./doc/examples/best_model.h5',
+                                load=True)
+        image_name = 'exampleimage.png'
+        image_path = './hardy/test/test_image/'
+        last_layer = cnn.feature_map(image_name, model, 4, 100,
                                      layer_num='last',
                                      image_path=image_path)
         assert isinstance(last_layer, np.ndarray),\
@@ -83,9 +83,9 @@ class TestSimulationTools(unittest.TestCase):
         assert last_layer[0][0] or last_layer[0][1] < 1,\
             'The output should be between 0 and 1'
 
-        cnn.feature_map(image_name, model, 2, 80, layer_num=None,
+        cnn.feature_map(image_name, model, 4, 100, layer_num=None,
                         image_path=image_path)
-        cnn.feature_map(image_name, model, 2, 80, layer_num=2,
+        cnn.feature_map(image_name, model, 4, 100, layer_num=2,
                         image_path=image_path)
 
         # test for array data
@@ -95,7 +95,7 @@ class TestSimulationTools(unittest.TestCase):
         image_array = img_to_array(image_load)
 
         last_layer_array = cnn.feature_map(
-            image_array, model, 2, 80, layer_num='last')
+            image_array, model, 4, 100, layer_num='last')
 
         assert isinstance(last_layer_array, np.ndarray),\
             'Returned output type must be numpy.ndarray'
@@ -108,9 +108,9 @@ class TestSimulationTools(unittest.TestCase):
         assert last_layer_array[0][0] or last_layer_array[0][1] < 1,\
             'The output should be between 0 and 1'
 
-        cnn.feature_map(image_array, model, 2, 80, layer_num=None)
+        cnn.feature_map(image_array, model, 4, 100, layer_num=None)
 
-        cnn.feature_map(image_array, model, 2, 80, layer_num=2)
+        cnn.feature_map(image_array, model, 4, 100, layer_num=2)
 
         # remove the images and folder just generated by the feature_map fnct.
         new_folder_path = "./report/"
