### 2021-12-17 3afcbe4
Added the feature maps documentations
