+++ doc/advanced.rst
@@ -47,3 +47,4 @@ Using Advanced Tools & Customization
     examples/using_data_reporting.ipynb
     examples/numpy_directory.ipynb
     examples/kfoldvalidation.ipynb
+    examples/featuremap.ipynb
