### 2021-12-17 e423724
fixed the notebook name error in advanced documentations
