### 2021-12-17 ea7b9c9
Added the docstrings for data wrapper in run_hardy
