+++ hardy/run_hardy.py
@@ -49,7 +49,9 @@ def hardy_main(  # Data and Config Paths
                     option to use images from arrays directly or save the
                     .png and use a directory iterator mode
     plot_format : str
-                  option for standard or RGB color gradient
+                  option for standard or RGB color gradient. If 'RGBrgb' it
+                  generates images in RGB format otherwise, it generates
+                  cartesian plots
     print_out : bool
                 option for printing out feedback on conputational time taken to
                 initialize the data and generate the images
@@ -59,7 +61,7 @@ def hardy_main(  # Data and Config Paths
     classifier : str
                   option cnn or tuner
     scale :  float
-             percentage fo the image to reduce its size to.
+             percentage for the image to reduce its size to.
     split : float
             the percentage of the learning set to use for the validation step
     target_size : tuple
@@ -87,6 +89,8 @@ def hardy_main(  # Data and Config Paths
     n_threads: int
                Number of cores used for parallel processing during the data
                transformation stage.
+    skiprows: int
+              Integer value representing the header size in the .csv files
 
     Function Calls  (see their related documentation)
     --------------
@@ -110,8 +114,8 @@ def hardy_main(  # Data and Config Paths
 
     Returns
     -------
-    * returns a folder containing subfolders, one for each trial run in the
-    hardy module. THese folders will containg a report on the run, as well as
+    returns a folder containing subfolders, one for each trial run in the
+    hardy module. These folders will containg a report on the run, as well as
     the best classifier model
 
     """
@@ -202,9 +206,52 @@ def data_wrapper(run_name=None, raw_datapath='./', tform_command_dict=None,
                  skiprows=0, scale=1.0):
     """
     Overall "One-Click" Wrapper to create the three "Keras Ready" Datasets
-        needed to train the model: "Training Set", "Validation Set" and
-        "Test Set", all in the same format which is created via the
-        Keras.Preprocessing.Data.Flow (<--- Not exact package/function)
+    needed to train the model: "Training Set", "Validation Set" and
+     "Test Set", all in the same format which is created via the
+    Keras.Preprocessing.Data.Flow (<--- Not exact package/function)
+
+    Parameters
+    ----------
+    run name: str
+              String value representing the transfomation name if called by
+              hardy_main() other wise can be any string
+    raw_datapath: str
+                   path to the raw .csv files containing the data to classify
+    tform_command_dict: dict
+                        dictionary containing the tranformation commands
+                        extracted from the transfomation configuration file.
+                        It can be generated by
+                        arbitrage.import_tform_config(tform_config_path)
+    classes : list
+              A list containing strings of the classes the data is divided in.
+              The class name represent the folder name the files are contained
+              in.
+    plot_format : str
+                  option for standard or RGB color gradient. If 'RGBrgb' it
+                  generates images in RGB format otherwise, it generates
+                  cartesian plots
+    iterator_mode : str
+                    option to use images from arrays directly or save the
+                    .png and use a directory iterator mode. If 'arrays',
+                    it generates the arrays. Other values generates images
+    print_out : bool
+                option for printing out feedback on conputational time taken to
+                initialize the data and generate the images
+    project_name : str
+                   name of the folder to be created for storing the results of
+                   the tuning or for generated array/images data
+    skiprows: int
+              Integer value representing the header size in the .csv files
+    scale :  float
+             percentage for the image to reduce its size to.
+
+    Returns
+    -------
+    array_data: .pkl
+                pickle file representing the images in the array format.
+                Stored in raw data path
+    images: .png
+            png images in case mode iterator mode is not 'arrays'
 
     """
     # replace the tform_commands with tform_command_dict in the arguments
