### 2022-01-10 #3829 @MariaPoliti
@andrewtarzia Thank you so much for the feedback and the review you provided for this paper. The remaining point on your checklist have been addressed:
* A new version of the manuscript has been uploaded on the repository with new figures and captions.
* Jupyter Notebooks/Google Collab notebooks have been uplaoded in the `doc/examples/` folder in the repository. These cover a variety of use cases and functionalities of the software.
* The documentation has been consolidated in the aformentioned folder and should now be more accessible and easy to find.

@gkthiruvathukal adn @alberto-battistel : Is there any update for this publication? When can we expect the review to be completed?

Thank you again all for your work and your feedback on this software!
