### 2022-01-26 EISy-as-Py/hardy#13 @alberto-battistel
[BUG] : Conda installation failed: problem with ResolvePackageNotFound: keras-tuner==1.0.1 
**Describe the bug**
Fail to install via conda.

**To Reproduce**
In terminal create a fresh environment:

`conda create --name joss_HARDy`

Activate the environment:

`conda activate joss_HARDy`

then run:

`conda install -c pozzorg hardy`

It gives:

>
> Collecting package metadata (current_repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
> Collecting package metadata (repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Solving environment: -
> Found conflicts! Looking for incompatible packages.
> This can take several minutes.  Press CTRL-C to abort.
> failed
>
> ResolvePackageNotFound:
>   - keras-tuner==1.0.1
>

**System information**
> OS: Manjaro Linux x86_64
> Host: 20AWS1J001 ThinkPad T440p
> Kernel: 5.4.173-1-MANJARO
> conda 4.11.0
