### 2022-01-26 EISy-as-Py/hardy#14 @alberto-battistel
[BUG] : GitHub repository installation failed: Permission denied
**Describe the bug**
The installation via GitHub repository failed because of Permission denied.

**To Reproduce**
Following the installation instruction.
In the terminal:
`git clone https://github.com/EISy-as-Py/hardy.git`

`cd hardy`

`git checkout v1.0`

`python setup.py install`

This gives:

> /usr/lib/python3.10/distutils/dist.py:274: UserWarning: Unknown distribution option: 'description_content_type'
>   warnings.warn(msg)
> running install
> /usr/lib/python3.10/site-packages/setuptools/command/install.py:34: SetuptoolsDeprecationWarning: setup.py install is deprecated. Use build and pip and other standards-based tools.
>   warnings.warn(
> /usr/lib/python3.10/site-packages/setuptools/command/easy_install.py:156: EasyInstallDeprecationWarning: easy_install command is deprecated. Use build and pip and other standards-based tools.
>   warnings.warn(
> error: can't create or remove files in install directory
>
> The following error occurred while trying to add or remove files in the
> installation directory:
>
>     [Errno 13] Permission denied: '/usr/lib/python3.10/site-packages/test-easy-install-21091.write-test'
>
> The installation directory you specified (via --install-dir, --prefix, or
> the distutils default setting) was:
>
>     /usr/lib/python3.10/site-packages/
>
> Perhaps your account does not have write access to this directory?  If the
> installation directory is a system-owned directory, you may need to sign in
> as the administrator or "root" account.  If you do not have administrative
> access to this machine, you may wish to choose a different installation
> directory, preferably one that is listed in your PYTHONPATH environment
> variable.
>
> For information on other options, you may wish to consult the
> documentation at:
>
>   https://setuptools.readthedocs.io/en/latest/deprecated/easy_install.html
>
> Please make the appropriate changes for your system and try again.
>

**Expected behavior**
The installation did not go through and one possible solution is to run the last step with administration rights or to change the installation directory.

**System info:**
OS: Manjaro Linux x86_64
Host: 20AWS1J001 ThinkPad T440p
Kernel: 5.4.173-1-MANJARO
