### 2022-01-26 EISy-as-Py/hardy#15 @alberto-battistel
[BUG] : Installation using evironment.yml failed
**Describe the bug**
The Installation using evironment.yml (Recommended) failed.

**To Reproduce**
Situation 1:
Download the 'evironment.yml' and run:

`conda install --name hardy --file environment.yml`

gives:

> EnvironmentLocationNotFound: Not a conda environment: /home/alberto/anaconda3/envs/hardy

Situation 2:
Clone the repository:

`git clone https://github.com/EISy-as-Py/hardy.git`

and run:
`cd hardy`
`conda install --name hardy --file environment.yml`

gives:

> EnvironmentLocationNotFound: Not a conda environment: /home/alberto/anaconda3/envs/hardy

Create the environment:

`conda create --name hardy`

and run:

`conda install --name hardy --file environment.yml`

still fails with:

> CondaValueError: could not parse 'name: hardy' in: environment.yml

**Expected behavior**
Maybe I am doing something fundamentally stupid, but I cannot succeed in installing this package.

**System information**

> OS: Manjaro Linux x86_64
> Host: 20AWS1J001 ThinkPad T440p
> Kernel: 5.4.173-1-MANJARO
> conda 4.11.0
