### 2022-01-26 #3829 @alberto-battistel
Hello @MariaPoliti,
I apologize for my very slow response.

I am trying installing the package. I tried all proposed methods, but they all fail.

**Installation via conda**
Installing the package via conda as suggested, but it failed. I opened an Issue with the description, see [https://github.com/EISy-as-Py/hardy/issues/13#issue-1114920637](url).

**Installation via GitHub repository**
I also tried following the GitHub repository, but is also failed. I opened another Issues with the details:
[https://github.com/EISy-as-Py/hardy/issues/14#issue-1114946713](url)

**Installation using evironment.yml**
Following [https://hardy.readthedocs.io/en/latest/installation.html#installation-using-evironment-yml-recommended](url) I tried the recommended way, but it also failed.
I opened another Issue with the details: [https://github.com/EISy-as-Py/hardy/issues/15#issue-1115195845](url)
