### 2022-01-26 EISy-as-Py/hardy#13 @amoeez
Hi @alberto-battistel,
When you create the environment, can you enforce python==3.7 into the conda argument, like the one below?

`conda create --name joss_HARDy python==3.7`
