### 2022-01-26 EISy-as-Py/hardy#15 @amoeez
That would be an issue while resolving the dependencies, though for this particular issue, the alternate is creating the environment using
`conda env create -f environment.yml`
after you clone the repository and change the folder in situation 2
