### 2022-01-26 EISy-as-Py/hardy#13 @alberto-battistel
Hi @amoeezuw,
yes,

`conda create --name joss_HARDy python==3.7`

works, but after

`conda activate joss_HARDy`

and

`conda install -c pozzorg hardy`

I get:

> Collecting package metadata (current_repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
> Collecting package metadata (repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Solving environment: \
> Found conflicts! Looking for incompatible packages.
> This can take several minutes.  Press CTRL-C to abort.
> failed
>
> UnsatisfiableError: The following specifications were found to be incompatible with each other:
>
> Output in format: Requested package -> Available versionsThe following specifications were found to be incompatible with your system:
>
>   - feature:/linux-64::__glibc==2.33=0
>   - feature:|@/linux-64::__glibc==2.33=0
>
> Your installed version is: 2.33
