### 2022-01-26 EISy-as-Py/hardy#13 @amoeez
Hi @alberto-battistel,

After you activate the environment, can you run either:

`pip install .` from the hardy directory or

`python3.7 setup.py install` instead of conda install.
