### 2022-01-28 EISy-as-Py/hardy#14 @amoeez
@alberto-battistel running the last command with administrative rights like `sudo python setup.py install` will resolve the issue.

However, through the information you provided above, I have noticed that your default python version is 3.10 which won't be compatible with HARDy
