### 2022-01-28 #3829 @alberto-battistel
Dear @MariaPoliti <https://github.com/MariaPoliti>,

I have read the manuscript. I used the version it was send by .@wedon in
the very beginning.

See in the attachment for a couple of comments I have for the text.

In  hardy/joss-paper/paper.md I see the .md file, but the captions for
the figures are missing. This was a point to improve mentioned by
@andrewtarzia <https://github.com/andrewtarzia>.

If you have a new version of the manuscript I will be happy to go
through it directly.

General comments on the manuscript

The manuscript is well-written and clear, but some details and
explanations are missing.

 1. Figure 1 should be explained more in detail.
 2. Figure 2 should also be explained more in detail. In particular, the
    Small Angle Scattering (SAS) data should be better described.
 3. It is not clear what is the RGB representation of the data.

I am going to flash a drive with a standard Ubuntu system and try the
installation one more time.
