### 2022-01-28 EISy-as-Py/hardy#14 @alberto-battistel
Yes, my default python version is 3.10.
I am going to flash a drive with an Ubuntu system and try one more time.
