### 2022-01-28 EISy-as-Py/hardy#15 @alberto-battistel
Ok, `conda env create -f environment.yml` works.
Python version in the environment is

> Python 3.7.12

However, after activating the environment and running

`conda install -c pozzorg hardy`

I get:

> Collecting package metadata (current_repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
> Collecting package metadata (repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Solving environment: \
> Found conflicts! Looking for incompatible packages.
> This can take several minutes.  Press CTRL-C to abort.
> failed
>
> UnsatisfiableError: The following specifications were found to be incompatible with each other:
>
> Output in format: Requested package -> Available versionsThe following specifications were found to be incompatible with your system:
>
>   - feature:/linux-64::__glibc==2.33=0
>   - feature:|@/linux-64::__glibc==2.33=0
>
> Your installed version is: 2.33
