### 2022-01-29 EISy-as-Py/hardy#15 @amoeez
Hi @alberto-battistel,

It appears conda is having conflicts with some already installed libraries on your OS. Although the version of dependencies appears the same, I'm unable to trace the root cause at the moment. I'll keep this issue open and meanwhile it is resolved, you should be able to install it either with `pip install .` or `python3.7 setup.py install` from the hardy directory.

Thank you for your patience.
