### 2022-01-29 EISy-as-Py/hardy#14 @amoeez
@alberto-battistel
We've some binder ready notebooks for testing of less computationally intensive functionalities, if you would like to use. Binder uses Ubuntu18.04 so it will save some of your time. The binder link is as follow:
https://mybinder.org/v2/gh/EISy-as-Py/hardy/v1.0

Following two notebooks are binder ready in `hardy/doc/examples`
`How_to_make_predictions_using_trained_model.ipynb`
`using_data_reporting.ipynb`

The rendering of these notebooks are available in the documentations:
https://hardy.readthedocs.io/en/latest/examples/How_to_make_predictions_using_trained_model.html
https://hardy.readthedocs.io/en/latest/examples/using_data_reporting.html

Thank you.
