### 2022-01-29 EISy-as-Py/hardy#13 @alberto-battistel
Ok, both `pip install .` and `python3.7 setup.py install` worked.
running `python -c "import hardy"` gives:

> Using TensorFlow backend.

running from the hardy folder:

`jupyter notebook`

and selecting `How_to_make_predictions_using_trained_model.ipynb` works as expected in both cases.

Thank you very much!
