### 2022-01-30 EISy-as-Py/hardy#15 @alberto-battistel
ok, `pip install .` works
