### 2022-02-04 #3829 @MariaPoliti
@alberto-battistel @andrewtarzia As mentioned in a previous message, the documentation has been updated and additional examples have been added- see `doc/examples` folder.
If the documentation is not to a satisfactory level, we can add more examples/explanations.

@alberto-battistel  the manuscript version on the repository is the final one. It seems like the captions are nto shown when opening the `.md` file on GitHub directly. If whedon does not generate a new version, you can generate it directly on this link: https://whedon.theoj.org/ by just pasting the repository's URL [ https://github.com/EISy-as-Py/hardy ].

Let us know if anything can be improved or need further attention.

Thank you all again for the feedback provided
