+++ setup.py
@@ -39,7 +39,7 @@ setup(name='hardy', version=1.0,
           "numpy==1.18.1",
           "pandas==1.0.3",
           "keras==2.3.1",
-          "tensorflow==2.2.0",
+          "tensorflow==2.5.3",
           "scikit-learn==0.23.1",
           "keras-tuner==1.0.1",
           "h5py==2.10.0",
