### 2022-02-11 EISy-as-Py/hardy#14 @alberto-battistel
I failed with the flashing of a drive with an Ubuntu system. It requires persistent storage on the drive and apparently I cannot get it to work.
However, I tried out the Binder and everything works as expected.

Thank you,
Alberto
