### 2022-02-22 #3829 @gkthiruvathukal
@MariaPoliti I think we're almost ready to go.

Can you scroll up to the checklist of items remaining to be addressed by @andrewtarzia? It looks like you addressed all of these comments, but I did not see any follow up from @andrewtarzia. It may not be needed but I would like you to confirm that you addressed all of the points there.

Once I have said confirmation, I can proceed with final checks and to recommend acceptance.
