### 2022-02-23 #3829 @andrewtarzia
Sorry about the delay. Having looked through the docs, examples and manuscript again, I am happy with it and recommend publication.
