### 2022-03-08 #3829 @MariaPoliti
@alberto-battistel There seems to be one task not checked off from your list. I believe we have addressed it, but I am just double checking with you.

@gkthiruvathukal What are the next steps after the review is completed? When can we expect for a final decision to be made about this publication?

Thank you agian all for your great feedback and your reviews
