### 2022-03-08 #3829 @gkthiruvathukal
@MariaPoliti, please do the following:

  [ ] Make a tagged release of your software, and list the version tag of the archived version here.
  [ ] Archive the reviewed software in Zenodo
  [ ] Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also  add the authors' ORCID.
  [ ] List the Zenodo DOI of the archived version here.

Once these are done, I'll be able to recommend acceptance.
