### 2022-03-09 #3829 @MariaPoliti
@gkthiruvathukal
Find the requested information below:

1. Release name: `HARDy v1.0.1 Release` ; Tag: `v1.0.1`
2. Zenodo Link: https://zenodo.org/record/6339767
3. Zenodo DOI: 10.5281/zenodo.6339767

Let me know if I have missed something
