### 2022-03-10 #3829 @gkthiruvathukal
@MariaPoliti  Thank you for your patience.
@andrewtarzia and @alberto-battistel Thank you for your great help on reviewing this submission.

I have recommended acceptance and look forward to seeing this work published in JOSS.
