### 2022-03-14 #3829 @kyleniemeyer
@MariaPoliti I made a few more minor formatting edits in this PR, can you merge it? https://github.com/EISy-as-Py/hardy/pull/17
